Resolves issue #98 - s3fs fails silently for bucket with upper case letters.
Research shows that many S3 tools do not support buckets whose
names contain uppercase characters -- s3fs appears to be no
different.  However, s3fs allows a bucket of this type to be
mounted even though access to it results in a "Input/output error"

Rather than put support in for these buckets, it's easier to
inform the user of the offending bucket name and die.